+++ pandora/matching_cost/sad_ssd.py
@@ -74,7 +74,7 @@ class SadSsd(matching_cost.AbstractMatchingCost):
         schema = {
             "matching_cost_method": And(str, lambda input: common.is_method(input, ["ssd", "sad"])),
             "window_size": And(int, lambda input: input > 0 and (input % 2) != 0),
-            "subpix": And(int, lambda input: input in (1, 2, 4)),
+            "subpix": And(int, lambda input: input > 0 and ((input % 2) == 0) or input == 1),
         }
 
         checker = Checker(schema)